requirements.txt
pytest == 8.0.2
selene >= 2.0.0rc7
allure-pytest==2.9.45